+++ settings.gradle
@@ -2,4 +2,3 @@ rootProject.name = 'survey'
 include 'survey-api'
 include 'survey-domain'
 include 'survey-common'
-include 'survey-security'
