--- survey-security/build.gradle
@@ -1,10 +1,0 @@
-plugins {
-    id 'survey.java-conventions'
-}
-
-dependencies {
-    implementation 'io.jsonwebtoken:jjwt-api:0.12.6'
-    implementation 'io.jsonwebtoken:jjwt-impl:0.12.6'
-    implementation 'io.jsonwebtoken:jjwt-jackson:0.12.6'
-    implementation project(':survey-domain')
-}
